# apps/backend/src/auth/auth.controller.ts
import {
  Controller,
  Post,
  Get,
  Body,
  UseGuards,
  Request,
  Res,
  HttpCode,
} from '@nestjs/common';
import {
  ApiTags,
  ApiOperation,
  ApiResponse,
  ApiBearerAuth,
} from '@nestjs/swagger';
import * as express from 'express';

import { JwtAuthGuard } from '../guards/jwt-auth.guard';
import { CreateSupportFeedbackDto } from '../user/dto/create-support-feedback.dto';
import { SupportFeedbackResponseDto } from '../user/dto/support-feedback-response.dto';
import { UserService } from '../user/user.service';

import { AuthService } from './auth.service';
import { ConfirmPairingDto } from './dto/confirm-pairing.dto';
import { RefreshQrDto } from './dto/refresh-qr.dto';
import { RequestPairingDto } from './dto/request-pairing.dto';
import { VerifyOtpDto } from './dto/verify-otp.dto';
import { VerifyPairingDto } from './dto/verify-pairing.dto';

const COOKIE_NAME = 'access_token';
const COOKIE_MAX_AGE = 7 * 24 * 60 * 60 * 1000; // 7 days

@ApiTags('auth')
@Controller('auth')
export class AuthController {
  constructor(
    private readonly authService: AuthService,
    private readonly userService: UserService,
  ) {}

  private setAuthCookie(res: express.Response, token: string): void {
    res.cookie(COOKIE_NAME, token, {
      httpOnly: true,
      secure: true,
      sameSite: 'none',
      maxAge: COOKIE_MAX_AGE,
      path: '/',
    });
  }

  @Post('request-pairing')
  @ApiOperation({
    summary: 'Request pairing code or check connection',
    description:
      'Request a pairing code to link WhatsApp account with the application. If the account is already connected, an OTP will be sent instead. For desktop devices, returns a QR scenario.',
  })
  @ApiResponse({
    status: 201,
    description: 'Pairing initiated successfully',
    schema: {
      type: 'object',
      properties: {
        scenario: {
          type: 'string',
          enum: [
            'otp',
            'pairing',
            'qr',
            'provisioning',
            'payment_required',
          ],
          example: 'otp',
        },
        code: { type: 'string', example: '12345678' },
        pairingToken: { type: 'string', example: 'token123...' },
        pricingUrl: { type: 'string', example: '/pricing' },
        message: {
          type: 'string',
          example: 'Un code de vérification a été envoyé',
        },
      },
    },
  })
  @ApiResponse({
    status: 400,
    description: 'Bad request - Invalid parameters',
  })
  async requestPairingCode(@Body() dto: RequestPairingDto) {
    return this.authService.requestPairingCode(
      dto.phoneNumber,
      dto.deviceType || 'mobile',
    );
  }

  @Post('request-qr')
  @ApiOperation({
    summary: 'Request QR code',
    description:
      'Request a QR code to link WhatsApp account with the application (desktop)',
  })
  @ApiResponse({
    status: 201,
    description: 'QR code generated successfully',
    schema: {
      type: 'object',
      properties: {
        qrCode: { type: 'string', example: '1@abc123...' },
        pairingToken: { type: 'string', example: 'token123...' },
        message: {
          type: 'string',
          example: 'Scannez le code QR avec WhatsApp',
        },
      },
    },
  })
  @ApiResponse({
    status: 400,
    description: 'Bad request - QR code not available',
  })
  async requestQRCode(
    @Body() dto: RequestPairingDto,
    @Res({ passthrough: true }) res: express.Response,
  ) {
    const result = await this.authService.requestCodeQR(dto.phoneNumber);
    if (result.accessToken) {
      this.setAuthCookie(res, result.accessToken);
    }
    return result;
  }

  @Post('refresh-qr')
  @ApiOperation({
    summary: 'Refresh QR code',
    description:
      'Request a new QR code when the current one expires. This will restart the connector to generate a fresh QR code.',
  })
  @ApiResponse({
    status: 201,
    description: 'QR code refreshed successfully',
    schema: {
      type: 'object',
      properties: {
        qrCode: { type: 'string', example: '1@abc123...' },
        pairingToken: { type: 'string', example: 'token123...' },
        message: {
          type: 'string',
          example: 'Nouveau code QR généré',
        },
      },
    },
  })
  @ApiResponse({
    status: 400,
    description: 'Bad request - Invalid pairing token',
  })
  async refreshQRCode(
    @Body() body: RefreshQrDto,
    @Res({ passthrough: true }) res: express.Response,
  ) {
    const result = await this.authService.refreshCodeQR(
      body.pairingToken,
      body.qrSessionToken,
    );
    if (result.accessToken) {
      this.setAuthCookie(res, result.accessToken);
    }
    return result;
  }

  @Post('verify-pairing')
  @ApiOperation({
    summary: 'Verify pairing success',
    description:
      'Verify that WhatsApp pairing was successful (called by webhook from connector)',
  })
  @ApiResponse({
    status: 201,
    description: 'Pairing verified successfully',
    schema: {
      type: 'object',
      properties: {
        accessToken: { type: 'string' },
        user: {
          type: 'object',
          properties: {
            id: { type: 'string' },
            phoneNumber: { type: 'string' },
            status: { type: 'string' },
            whatsappProfile: { type: 'object' },
          },
        },
      },
    },
  })
  @ApiResponse({
    status: 404,
    description: 'User not found',
  })
  async verifyPairing(
    @Body() dto: VerifyPairingDto,
    @Res({ passthrough: true }) res: express.Response,
  ) {
    const result = await this.authService.verifyPairingSuccess(
      dto.phoneNumber,
      dto.whatsappProfile,
    );
    if (result.accessToken) {
      this.setAuthCookie(res, result.accessToken);
    }
    return result;
  }

  @Post('confirm-pairing')
  @ApiOperation({
    summary: 'Confirm pairing completion or verify OTP',
    description:
      'Called by frontend to confirm pairing (new device) or verify OTP code (existing connection). The endpoint automatically detects which scenario to handle based on the presence of an OTP in cache.',
  })
  @ApiResponse({
    status: 201,
    description: 'Pairing/OTP confirmed successfully',
    schema: {
      type: 'object',
      properties: {
        accessToken: { type: 'string' },
        user: {
          type: 'object',
          properties: {
            id: { type: 'string' },
            phoneNumber: { type: 'string' },
            status: { type: 'string' },
            whatsappProfile: { type: 'object' },
          },
        },
      },
    },
  })
  @ApiResponse({
    status: 401,
    description: 'Invalid or expired pairing token / Invalid OTP code',
  })
  @ApiResponse({
    status: 400,
    description:
      'WhatsApp connection not yet confirmed / OTP code required but not provided',
  })
  async confirmPairing(
    @Body() dto: ConfirmPairingDto,
    @Res({ passthrough: true }) res: express.Response,
  ) {
    const result = await this.authService.confirmPairing(
      dto.pairingToken,
      dto.otpCode,
    );
    if (result.accessToken) {
      this.setAuthCookie(res, result.accessToken);
    }
    return result;
  }

  @Post('verify-otp')
  @ApiOperation({
    summary: 'Verify OTP code',
    description: 'Verify OTP code and complete login',
  })
  @ApiResponse({
    status: 201,
    description: 'OTP verified successfully',
    schema: {
      type: 'object',
      properties: {
        accessToken: { type: 'string' },
        user: {
          type: 'object',
          properties: {
            id: { type: 'string' },
            phoneNumber: { type: 'string' },
            status: { type: 'string' },
            whatsappProfile: { type: 'object' },
            lastLoginAt: { type: 'string', format: 'date-time' },
          },
        },
      },
    },
  })
  @ApiResponse({
    status: 401,
    description: 'Invalid or expired OTP',
  })
  async verifyOtp(
    @Body() dto: VerifyOtpDto,
    @Res({ passthrough: true }) res: express.Response,
  ) {
    const result = await this.authService.verifyOTP(dto.phoneNumber, dto.code);
    if (result.accessToken) {
      this.setAuthCookie(res, result.accessToken);
    }
    return result;
  }

  @Post('support-feedback')
  @HttpCode(201)
  @ApiOperation({ summary: 'Submit public support feedback' })
  @ApiResponse({
    status: 201,
    description: 'Support feedback submitted successfully',
    type: SupportFeedbackResponseDto,
  })
  @ApiResponse({ status: 400, description: 'Invalid feedback payload' })
  @ApiResponse({
    status: 500,
    description: 'Support feedback relay failed',
  })
  async submitSupportFeedback(
    @Body() createSupportFeedbackDto: CreateSupportFeedbackDto,
  ): Promise<SupportFeedbackResponseDto> {
    return this.userService.submitSupportFeedback(
      null,
      createSupportFeedbackDto,
    );
  }

  @Get('me')
  @UseGuards(JwtAuthGuard)
  @ApiBearerAuth()
  @ApiOperation({
    summary: 'Get current user',
    description: 'Get the currently authenticated user information',
  })
  @ApiResponse({
    status: 200,
    description: 'Current user information',
    schema: {
      type: 'object',
      properties: {
        id: { type: 'string' },
        phoneNumber: { type: 'string' },
        status: { type: 'string' },
        whatsappProfile: { type: 'object' },
        googleContacts: {
          type: 'object',
          properties: {
            connected: { type: 'boolean' },
            contactsCount: { type: 'number' },
          },
        },
      },
    },
  })
  @ApiResponse({
    status: 401,
    description: 'Unauthorized',
  })
  async getMe(@Request() req) {
    return req.user;
  }

  @Get('/debug-sentry')
  getError() {
    // if we are after 20 / 04 / 2026 do nothing
    if (new Date() > new Date('2026-04-20')) {
      return { message: 'Sentry test endpoint expired' };
    }
    throw new Error('My first Sentry error!');
  }

  @Post('logout')
  @HttpCode(200)
  @ApiOperation({
    summary: 'Logout',
    description: 'Clear the auth cookie to log out',
  })
  @ApiResponse({
    status: 200,
    description: 'Logged out successfully',
  })
  logout(@Res({ passthrough: true }) res: express.Response) {
    res.clearCookie(COOKIE_NAME, {
      httpOnly: true,
      secure: true,
      sameSite: 'none',
      path: '/',
    });
    return { message: 'Logged out' };
  }
}