# apps/backend/swagger-output/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/request-pairing": {
      "post": {
        "operationId": "AuthController_requestPairingCode",
        "summary": "Request pairing code or check connection",
        "description": "Request a pairing code to link WhatsApp account with the application. If the account is already connected, an OTP will be sent instead. For desktop devices, returns a QR scenario.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPairingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pairing initiated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "scenario": {
                      "type": "string",
                      "enum": [
                        "otp",
                        "pairing",
                        "qr",
                        "provisioning",
                        "payment_required"
                      ],
                      "example": "otp"
                    },
                    "code": {
                      "type": "string",
                      "example": "12345678"
                    },
                    "pairingToken": {
                      "type": "string",
                      "example": "token123..."
                    },
                    "pricingUrl": {
                      "type": "string",
                      "example": "/pricing"
                    },
                    "message": {
                      "type": "string",
                      "example": "Un code de vérification a été envoyé"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid parameters"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/request-qr": {
      "post": {
        "operationId": "AuthController_requestQRCode",
        "summary": "Request QR code",
        "description": "Request a QR code to link WhatsApp account with the application (desktop)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPairingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "QR code generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "qrCode": {
                      "type": "string",
                      "example": "1@abc123..."
                    },
                    "pairingToken": {
                      "type": "string",
                      "example": "token123..."
                    },
                    "message": {
                      "type": "string",
                      "example": "Scannez le code QR avec WhatsApp"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - QR code not available"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh-qr": {
      "post": {
        "operationId": "AuthController_refreshQRCode",
        "summary": "Refresh QR code",
        "description": "Request a new QR code when the current one expires. This will restart the connector to generate a fresh QR code.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshQrDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "QR code refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "qrCode": {
                      "type": "string",
                      "example": "1@abc123..."
                    },
                    "pairingToken": {
                      "type": "string",
                      "example": "token123..."
                    },
                    "message": {
                      "type": "string",
                      "example": "Nouveau code QR généré"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid pairing token"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/verify-pairing": {
      "post": {
        "operationId": "AuthController_verifyPairing",
        "summary": "Verify pairing success",
        "description": "Verify that WhatsApp pairing was successful (called by webhook from connector)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPairingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pairing verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "phoneNumber": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        },
                        "whatsappProfile": {
                          "type": "object"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/confirm-pairing": {
      "post": {
        "operationId": "AuthController_confirmPairing",
        "summary": "Confirm pairing completion or verify OTP",
        "description": "Called by frontend to confirm pairing (new device) or verify OTP code (existing connection). The endpoint automatically detects which scenario to handle based on the presence of an OTP in cache.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmPairingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pairing/OTP confirmed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "phoneNumber": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        },
                        "whatsappProfile": {
                          "type": "object"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "WhatsApp connection not yet confirmed / OTP code required but not provided"
          },
          "401": {
            "description": "Invalid or expired pairing token / Invalid OTP code"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/verify-otp": {
      "post": {
        "operationId": "AuthController_verifyOtp",
        "summary": "Verify OTP code",
        "description": "Verify OTP code and complete login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OTP verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "phoneNumber": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        },
                        "whatsappProfile": {
                          "type": "object"
                        },
                        "lastLoginAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired OTP"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/support-feedback": {
      "post": {
        "operationId": "AuthController_submitSupportFeedback",
        "summary": "Submit public support feedback",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupportFeedbackDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Support feedback submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportFeedbackResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid feedback payload"
          },
          "500": {
            "description": "Support feedback relay failed"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_getMe",
        "summary": "Get current user",
        "description": "Get the currently authenticated user information",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "phoneNumber": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "whatsappProfile": {
                      "type": "object"
                    },
                    "googleContacts": {
                      "type": "object",
                      "properties": {
                        "connected": {
                          "type": "boolean"
                        },
                        "contactsCount": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Monorepo API",
    "description": "Interractive documentation for the Monorepo API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RequestPairingDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "description": "Phone number in E.164 format",
            "example": "+15551234567"
          },
          "deviceType": {
            "type": "string",
            "description": "Device type (mobile or desktop)",
            "example": "mobile",
            "enum": [
              "mobile",
              "desktop"
            ]
          }
        },
        "required": [
          "phoneNumber"
        ]
      },
      "RefreshQrDto": {
        "type": "object",
        "properties": {
          "pairingToken": {
            "type": "string",
            "description": "Pairing token associated with the QR session",
            "example": "token123"
          },
          "qrSessionToken": {
            "type": "string",
            "description": "JWT de session QR signé côté backend, contient le numéro et le pairing token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "pairingToken",
          "qrSessionToken"
        ]
      },
      "VerifyPairingDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "description": "Phone number in E.164 format",
            "example": "+15551234567"
          },
          "whatsappProfile": {
            "type": "object",
            "description": "WhatsApp profile information",
            "example": {
              "name": "John Doe",
              "picture": "https://..."
            }
          }
        },
        "required": [
          "phoneNumber",
          "whatsappProfile"
        ]
      },
      "ConfirmPairingDto": {
        "type": "object",
        "properties": {
          "pairingToken": {
            "type": "string",
            "description": "The pairing token received when requesting the pairing code",
            "example": "a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6q7r8s9t0u1v2w3x4y5z6a7b8c9d0e1f2"
          },
          "otpCode": {
            "type": "string",
            "description": "OTP code if this is an OTP login scenario (required when scenario is \"otp\")",
            "example": "123456"
          }
        },
        "required": [
          "pairingToken"
        ]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "description": "Phone number in E.164 format",
            "example": "+15551234567"
          },
          "code": {
            "type": "string",
            "description": "6-digit OTP code",
            "example": "123456"
          }
        },
        "required": [
          "phoneNumber",
          "code"
        ]
      },
      "SupportFeedbackContextDto": {
        "type": "object",
        "properties": {
          "appArea": {
            "type": "string",
            "description": "Application area"
          },
          "currentPlan": {
            "type": "string",
            "description": "Detected plan label"
          },
          "route": {
            "type": "string",
            "description": "Current route pathname"
          },
          "timezone": {
            "type": "string",
            "description": "User timezone"
          },
          "url": {
            "type": "string",
            "description": "Current page absolute URL"
          },
          "userId": {
            "type": "string",
            "description": "User identifier from frontend context"
          },
          "contextScore": {
            "type": "string",
            "description": "Context score value"
          }
        }
      },
      "SupportFeedbackSentryDto": {
        "type": "object",
        "properties": {
          "dsn": {
            "type": "string",
            "description": "Public Sentry DSN used for support feedback",
            "example": "https://549db520247ca4aa7f69f7e3eb5775f6@o1063428.ingest.us.sentry.io/4511042903539712"
          },
          "environment": {
            "type": "string",
            "description": "Frontend Sentry environment"
          },
          "release": {
            "type": "string",
            "description": "Frontend release identifier"
          }
        },
        "required": [
          "dsn"
        ]
      },
      "CreateSupportFeedbackDto": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "description": "Support request category",
            "example": "bug"
          },
          "email": {
            "type": "string",
            "description": "Reply email used in Sentry feedback",
            "example": "client@example.com"
          },
          "message": {
            "type": "string",
            "description": "Support message body",
            "example": "Le formulaire support reste bloque apres validation."
          },
          "name": {
            "type": "string",
            "description": "Display name sent to support",
            "example": "Jean Dupont"
          },
          "subject": {
            "type": "string",
            "description": "Optional support subject",
            "example": "Bug sur la page Support"
          },
          "context": {
            "description": "Additional non-sensitive frontend context",
            "allOf": [
              {
                "$ref": "#/components/schemas/SupportFeedbackContextDto"
              }
            ]
          },
          "sentry": {
            "description": "Sentry relay configuration",
            "allOf": [
              {
                "$ref": "#/components/schemas/SupportFeedbackSentryDto"
              }
            ]
          }
        },
        "required": [
          "category",
          "email",
          "message",
          "name",
          "sentry"
        ]
      },
      "SupportFeedbackResponseDto": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "description": "Generated Sentry feedback event identifier",
            "example": "4f3c0f5d5d8748f394e8d4c7a1b23456"
          }
        },
        "required": [
          "eventId"
        ]
      },
...
    }
  }
}